% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_supplycurve.R
\name{calc_supplycurve}
\alias{calc_supplycurve}
\title{Calculate points of fitted curve for plotting using fit coefficients}
\usage{
calc_supplycurve(data_in, fitcoef, myform, ylimit = "common")
}
\arguments{
\item{data_in}{MAgPIE object containing the same data that was used to calculate the fitcoefficients.}

\item{fitcoef}{MAgPIE object containing the fitcoefficients (output of \code{\link{calculate_fit}})}

\item{myform}{Function that was fitted and is used here to calculate curve}

\item{ylimit}{Choose the method to calculate the upper limit for the y values. Options: "individual", the maximal value up to which 
the y values of the supplycurve are calculated is done individualy for each reagion and year. "common" (default), the upper limit of  
the y values of the supplycurve for all regions and years is the same and is the maximum of the raw data. This is useful for plotting,
because it expands "short" supplycurves beyond the maximal demand of the respective region and year and makes them better comparable.}
}
\value{
MAgPIE object containing the points on the curve.
}
\description{
This function uses the raw data on which you performed the fit (\code{\link{calculate_fit}}) and the resulting 
fit coefficients to calculate a couple of points on the fitted curve that can be used to plot the curve (\code{\link{plot_curve}}).
}
\seealso{
\code{\link{calculate_fit}} \code{\link{plot_curve}}
}
\author{
David Klein
}
