% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fit.R
\name{calculate_fit}
\alias{calculate_fit}
\title{Calculate fit by minimizing sum of error squares}
\usage{
calculate_fit(data, initial_values = c(1, 1, 1), form, ...)
}
\arguments{
\item{data}{MAgPIE object with data to fit.}

\item{initial_values}{Initial values for the parameters to be optimized over.}

\item{form}{Fit function for which least squares will be calaculated using the \code{optim} function.}

\item{...}{Arguments passed on to the \code{optim} function. Useful to define bounds on fit coefficients.}
}
\value{
MAgPIE object with fit coefficients a, b, and d
}
\description{
This is an internal function of the remulator package. It fits a curve to given data. For each region and year it minimizes the function \code{sum((y - (a + b * x ^d))^2}
x and y are the data provided by the user, and a, b, and d are the coefficients to be found. So, the fit is \code{y = a+b*x^d}.
}
\author{
David Klein
}
