% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{emulator}
\alias{emulator}
\title{Calculate and plot fit.}
\usage{
emulator(
  data,
  name_x,
  name_y,
  name_modelstat = NULL,
  treat_as_feasible = c(2, 7),
  userfun = function(param, x) return(param[[1]] + param[[2]] * x^param[[3]]),
  initial_values = c(0, 0, 1),
  outlier_range = 1.5,
  n_suff = 1,
  fill = FALSE,
  output_path = "emulator",
  fitname = "linear",
  create_pdf = TRUE,
  ...
)
}
\arguments{
\item{data}{MAgPIE object containing at least two variables and the modelstatus}

\item{name_x}{Name of the variable in \code{data} that will be treated as x in the fit}

\item{name_y}{Name of the variable in \code{data} that will be treated as y in the fit}

\item{name_modelstat}{Name of the variable that contains the modelstatus}

\item{treat_as_feasible}{GAMS model status codes that will be regarded feasible. 
See \url{https://www.gams.com/24.8/docs/userguides/mccarl/modelstat_tmodstat.htm}}

\item{userfun}{Function to fit. User can provide a functional form using the following 
syntax: \code{function(param,x)return(param[[1]] + param[[2]] * x ^param[[3]])}. This function is the default.}

\item{initial_values}{Vector with initial values of the fit coefficients.}

\item{outlier_range}{Before the actual fit a linear pre-fit is performed. Based on their distance to this 
pre-fit the data points are allocated to quartiles. A data point is considered an outlier if it is more 
than \code{outlier_range} times of the interquartile range away from the the upper or lower end of the 
interquartile range \url{http://colingorrie.github.io/outlier-detection.html}.}

\item{n_suff}{Minial number (default=1) of data points in a specific year and region that 
will be regarded as sufficient to perform a fit.
If the number of available data points is less no fit will be generated for this year and region.}

\item{fill}{Logical (default=FALSE) indicating whether data will be copied from subsequent year if in the 
current year not enough data points are avaialbe.}

\item{output_path}{Path to save the output to}

\item{fitname}{Name that describes the fit (default: linear) and will be used for naming the output folders.}

\item{create_pdf}{Logical indicating whether a pdf should be produced that compiles all figures.}

\item{...}{Arguments passed on to the \code{optim} function in \code{calcualte_fit}. Useful to define bounds on fit coefficients.}
}
\value{
MAgPIE object containning fit coefficients
}
\description{
This function fits a function of the form \code{y = a + b * x ^ d} for the given data.
}
\author{
David Klein
}
