% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator_runs_complete.R
\name{emulator_runs_complete}
\alias{emulator_runs_complete}
\title{Checks whether set of given runs is complete.}
\usage{
emulator_runs_complete(list_of_directories, runnumbers = 1:73)
}
\arguments{
\item{list_of_directories}{Vector of strings containing path names. Must have "-xy" at the end of each name, with xy beinga one or two digit number.}

\item{runnumbers}{Vector of integers providing the number that have to be present in \code{list_of_directories}.}
}
\value{
Logical. TRUE if set of numbers at the end of the path names is idential to \code{runnumbers}.
}
\description{
Checks whether set of given runs is complete.
}
\author{
David Klein
}
