% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_years.R
\name{fill_missing_years}
\alias{fill_missing_years}
\title{Find fit coefficients for years where no fit could be calculated}
\usage{
fill_missing_years(fitcoef, nodata, method = 1)
}
\arguments{
\item{fitcoef}{MAgPIE object containing the fit coefficients}

\item{nodata}{MAgPIE object of the same shape as \code{fitcoef} containing only logicals that indicate where data is not available (TRUE)}

\item{method}{Choose method that finds fit coefficients for years with no fit. Currently there is only one method availalbe that in a first step
takes the fit from the subsequent year. If all subsequent years have no fit it takes the fit from the preceding year.}
}
\value{
Magpie object with the updated fit coefficients. Has an attribute attached providing the years missing fits were taken from.
}
\description{
Find fit coefficients for years where no fit could be calculated
}
\author{
David Klein
}
