% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize_least_squares.R
\name{minimize_least_squares}
\alias{minimize_least_squares}
\title{Minimizes the sum of the error squares}
\usage{
minimize_least_squares(dat, initial_values, userform, ...)
}
\arguments{
\item{dat}{Array with data points to fit containing one column with sample numbers, one column with x values, and one column with y values}

\item{initial_values}{Initial values for the parameters to be optimized over.}

\item{userform}{Fit function for which least squares will be calaculated using the \code{optim} function.}

\item{...}{Arguments passed on to the \code{optim} function. Useful to define bounds on fit coefficients.}
}
\value{
list containing fit coefficients and messages
}
\description{
This is an internal function of the remulator package. It fits a user defined function to data minimizing the sum of the error squares.
}
\author{
David Klein
}
