% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mute_duplicated.R
\name{mute_duplicated}
\alias{mute_duplicated}
\title{Set duplicated samples to NA}
\usage{
mute_duplicated(data)
}
\arguments{
\item{data}{MAgPIE object containing the samples to remove the duplicates from.}
}
\value{
Magpie object with duplicated samples set to NA.
}
\description{
This function sets duplicated samples (per region, year, model, scenario) to NA. This is
useful to clean your data before fitting.
}
\author{
David Klein
}
