% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mute_infes.R
\name{mute_infes}
\alias{mute_infes}
\title{Set data of infeasible years and their successors to NA}
\usage{
mute_infes(data, name = "Modelstatus (-)", feasible = 2)
}
\arguments{
\item{data}{MAgPIE object containing the results of a MAgPIE run and the modelstatus.}

\item{name}{String providing the name of the variable that holds the modelstatus.}

\item{feasible}{Integer vector defining which modelstatus will be treated as feasible.}
}
\value{
Magpie object with either filtered model data or the modelstatus.
}
\description{
This function sets data of infeasible years and all years after the first infeasible
year to NA because results after an infeasible year are not meaningful. The modelstatus
has to be provided as one of the variables in the data. This function will look for the
modelstatus variable with the name given in \code{name}. If \code{return_infes} is TRUE
this function return the matrix with infeasible years instead of the data.
}
\author{
David Klein
}
