% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mute_insufficient.R
\name{mute_insufficient}
\alias{mute_insufficient}
\title{In years and regions with only an unsufficient number of data points set all data points to NA}
\usage{
mute_insufficient(data, n_suff)
}
\arguments{
\item{data}{MAgPIE object containing the samples to remove the duplicates from.}

\item{n_suff}{Minial number (default=1) of data points that will be regarded as sufficient to perform a fit.}
}
\value{
Magpie object with duplicated samples set to NA.
}
\description{
This function counts data points (per region, year, scenario) and sets them to NA if the count
is less than given in \code{n_suff}. This is useful to clean your data before fitting. No fit will
be calculated where no data is available.
}
\author{
David Klein
}
