% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mute_outliers.R
\name{mute_outliers}
\alias{mute_outliers}
\title{Set outlier samples to NA}
\usage{
mute_outliers(data, range = 1.5)
}
\arguments{
\item{data}{MAgPIE object containing the samples to remove the duplicates from.}

\item{range}{Multiplied with the interquartile range this is the maximal distance a data point may have to not be considered an outlier.}
}
\value{
Magpie object with duplicated samples set to NA.
}
\description{
This function sets outlier samples (per region, year, model, scenario) to NA. This is useful to 
clean your data before fitting. A data point is considered an outlier if it is more than \code{range} times
of the interquartile range away from the the upper or lower end of the interquartile range
\url{http://colingorrie.github.io/outlier-detection.html}
}
\seealso{
\code{\link[graphics]{boxplot}}
}
\author{
David Klein
}
