% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_supplycurves.R
\name{plot_compare_supplycurves}
\alias{plot_compare_supplycurves}
\title{Calculate and plot multiple supplycurves to one plot}
\usage{
plot_compare_supplycurves(folders, pdfname = NULL)
}
\arguments{
\item{folders}{Vector giving the paths to the fitted data (usually within the 'output/emulator' folder)}

\item{pdfname}{If you want the figures to be compiled in a pdf provide the name of the file here}
}
\description{
This function calculates supplycurves of multiple scenarios and plots them into one figure to compare them.
}
\seealso{
\code{\link{emulator}}
}
\author{
David Klein
}
