% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curve.R
\name{plot_curve}
\alias{plot_curve}
\title{Plot fitted curve to png files and additionally compile them in a pdf file.}
\usage{
plot_curve(
  raw,
  supplycurve_commonY,
  supplycurve_indiviY,
  infes,
  emu_path = "emulator",
  fitname = "linear",
  create_pdf = TRUE
)
}
\arguments{
\item{raw}{MAgPIE object containing the same raw data that was used to calculate the fitcoefficients.}

\item{supplycurve_commonY}{MAgPIE object containing the points of the curve (with common y limit) (output of \code{\link{calc_supplycurve}})}

\item{supplycurve_indiviY}{MAgPIE object containing the points of the curve (with individual y limit) (output of \code{\link{calc_supplycurve}})}

\item{infes}{MAgPIE object containing the modelstatus (optional output of \code{\link{mute_infes}})}

\item{emu_path}{Name of the folder the figures and pdf will be saved to.}

\item{fitname}{Name that describes the fit (default: linear) and will be used for naming the output folders.}

\item{create_pdf}{Logical indicating whether a pdf should be produced that compiles all figures.}
}
\description{
This function produces plot showing the raw data and the fittet curves. It saves it to a folder named after the
scenario and it produces a pdf containing all the figures.
}
\seealso{
\code{\link{calc_supplycurve}} \code{\link{mute_infes}}
}
\author{
David Klein
}
