% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_combine.R
\name{read_and_combine}
\alias{read_and_combine}
\title{Read multiple report mif files from REMIND or MAgPIE and combine into single magpie object}
\usage{
read_and_combine(list_of_directories, outfile = NULL)
}
\arguments{
\item{list_of_directories}{Vector of strings providing the folder names from which reports will be read.}

\item{outfile}{Path to an output file which the output of the function can optionally be stored to.}
}
\description{
This function reads model output from mif files from the provided list of folders. It also reads the GAMS
modelstatus from the fulldata.gdx and appends it to the report. If more than one folder is
given it combines the contents of the multiple reports into a single magpie object and returns it.
If the name of an output file is provided the object is additionally stored in this file. In each of the
folders there must be a report_scenname.mif, a fulldata.gdx, a spatial_header.rda, and config.Rdata.
}
\author{
David Klein
}
