% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_flat_fits.R
\name{replace_flat_fits}
\alias{replace_flat_fits}
\title{Replace fits in years where they are flat with fits from other years}
\usage{
replace_flat_fits(
  path_to_postfit_rdata,
  flat = NULL,
  threshold = 0.01,
  plot = FALSE
)
}
\arguments{
\item{path_to_postfit_rdata}{Path to the Rdata file that contains raw data,
fitted data and fitcoefficients saved by the emulator function}

\item{flat}{If you want this function to replace further fits provide
their region and year here using a vector of the form
c("LAM:2020,2025", "IND:2005")}

\item{threshold}{Defines the value of the slope (b) below which fits are considered flat}

\item{plot}{logical. If TRUE supply curves will be saved to png files}
}
\description{
This function considers fits with a slope less than defined in threshold flat and tries to
replace them with non-flat fits from other years.
}
\seealso{
\code{\link{emulator}} \code{\link{fill_missing_years}}
}
\author{
David Klein
}
